[CI] Flatten hybrid language generation tests into whole files
Move the Hybrid lane out of test_hybrid.py into whole files placed
directly under tests/models/language/generation/hybrid/ (flat, no
subdirectories), with the GDN sleep/wake test moved whole from
generation/. The test_models body, model lists and APC helpers live in
the shared, non-collected _hybrid_models.py (leading underscore),
imported by each file with unchanged parametrize stacks so display IDs
and the hybrid_model selection are unchanged;
tests/tools/test_language_layout.py proves the helper is never
collected.

The granite-4.0-tiny-preview case stays a lone file at generation/
level (test_granite_4_hybrid.py): its H200 output does not match the
HF reference, so production keeps it on a dedicated L4 job - measured
evidence: validation build 88231, where the case failed the HF vs vLLM
logprob comparison when run inside the hybrid shards on H200.

The layout is flat by design: CI will measure per-file timings from
this layout and any future split of the lane will be designed from
that data. The language area YAML is untouched; the existing
-m hybrid_model commands select the same case multiset over the new
tree.
