+++ tests/models/language/generation/hybrid/test_models_falcon_lfm_qwen.py
@@ -1,0 +1,37 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+
+import pytest
+
+from tests.models.language.generation._hybrid_models import _test_models
+
+pytestmark = pytest.mark.hybrid_model
+
+MODELS = [
+    "tiiuae/Falcon-H1-0.5B-Base",
+    "LiquidAI/LFM2-1.2B",
+    "tiny-random/qwen3-next-moe",
+]
+
+
+@pytest.mark.parametrize("model", MODELS)
+@pytest.mark.parametrize("max_tokens", [64])
+@pytest.mark.parametrize("num_logprobs", [5])
+def test_models(
+    hf_runner,
+    vllm_runner,
+    example_prompts,
+    monkeypatch,
+    model: str,
+    max_tokens: int,
+    num_logprobs: int,
+) -> None:
+    _test_models(
+        hf_runner,
+        vllm_runner,
+        example_prompts,
+        monkeypatch,
+        model,
+        max_tokens,
+        num_logprobs,
+    )
